+++ lib/legion/api/middleware/request_logger.rb
@@ -9,15 +9,18 @@ module Legion
         end
 
         def call(env)
+          method_path = "#{env['REQUEST_METHOD']} #{env['PATH_INFO']}"
+          Legion::Logging.info "[api][request-start] #{method_path}"
           start = ::Process.clock_gettime(::Process::CLOCK_MONOTONIC)
           status, headers, body = @app.call(env)
           duration = ((::Process.clock_gettime(::Process::CLOCK_MONOTONIC) - start) * 1000).round(2)
 
-          Legion::Logging.info "[api] #{env['REQUEST_METHOD']} #{env['PATH_INFO']} #{status} #{duration}ms"
+          level = duration > 5000 ? :warn : :info
+          Legion::Logging.send(level, "[api] #{method_path} #{status} #{duration}ms")
           [status, headers, body]
         rescue StandardError => e
           duration = ((::Process.clock_gettime(::Process::CLOCK_MONOTONIC) - start) * 1000).round(2)
-          Legion::Logging.error "[api] #{env['REQUEST_METHOD']} #{env['PATH_INFO']} 500 #{duration}ms - #{e.message}"
+          Legion::Logging.error "[api] #{method_path} 500 #{duration}ms - #{e.message}"
           raise
         end
       end